+++ src/app/models/hris/employee-qualifications.interface.ts
@@ -1,10 +1,9 @@
 export interface EmployeeQualifications {
     id : number;
-    employeeId?: number;
-    highestQualification?: number ,
-    school?: string ,
-    degree?: string ,
-    fieldOfStudy?: string,
-    year?: Date["getUTCDate"],
-    nqfLevel?: number,
+    highestQualification: string | null,
+    school: string | null,
+    degree: string | null,
+    fieldOfStudy: string | null,
+    yearObtained: string | null,
+    nqfLevel: string | null,
 }
